Refactor: ShareProfile functionality was moved into its own comp; Feat: Added friendscard modal to see usercodes; Migration: Made adjustments to the public.friends rpc function to get all cols from the respective table making it possible to use the generated supabase types to make the rpc typesafe allowing for better dynamic behavior and scalability
