+++ supabase/seeds/public.notifications.sql
@@ -1,1 +1,1 @@
-INSERT INTO "public"."notifications" ("id", "created_at", "type", "recipient", "marked_as_read", "raw_notification_meta_data", "triggered_by") VALUES ('1', '2025-01-29 22:39:08.84226+00', 'new_follower', '98729206-98e6-49d4-a19e-0918c3a78399', 'false', null, '8071ec28-fe6f-4be6-a6e0-60e0475073ba'), ('2', '2025-01-29 22:48:25.274407+00', 'new_follower', '8071ec28-fe6f-4be6-a6e0-60e0475073ba', 'false', null, 'cbce80a3-5cb4-4e6b-97a7-3fa7dab4ce1b'), ('3', '2025-01-29 23:26:38.361949+00', 'code_interaction', '98729206-98e6-49d4-a19e-0918c3a78399', 'false', null, '8071ec28-fe6f-4be6-a6e0-60e0475073ba');
+INSERT INTO "public"."notifications" ("id", "created_at", "type", "recipient", "marked_as_read", "raw_notification_meta_data", "triggered_by") VALUES ('1', '2025-01-29 22:39:08.84226+00', 'new_friend', '98729206-98e6-49d4-a19e-0918c3a78399', 'false', null, '8071ec28-fe6f-4be6-a6e0-60e0475073ba'), ('2', '2025-01-29 22:48:25.274407+00', 'new_friend', '8071ec28-fe6f-4be6-a6e0-60e0475073ba', 'false', null, 'cbce80a3-5cb4-4e6b-97a7-3fa7dab4ce1b'), ('3', '2025-01-29 23:26:38.361949+00', 'code_interaction', '98729206-98e6-49d4-a19e-0918c3a78399', 'false', null, '8071ec28-fe6f-4be6-a6e0-60e0475073ba');
