+++ types/general.types.ts
@@ -31,22 +31,8 @@ export type UserCodeWithRelations = Tables<"user_codes"> & {
 
 
 //Types for rpc function to get friends with their respective user codes
-export type UserCode = {
-  id: number;
-  referral_value: string | null;
-  company: {
-    id: number;
-    name: string;
-    logo_url: string;
-  };
-};
-
 export type FriendWithCodes = {
   created_at: string; // Timestamp from the friends table
-  profile: {
-    id: string;
-    avatar_url: string;
-    user_name: string;
-  }
-  user_codes: UserCode[];
+  profile: Tables<"profiles">
+  user_codes: UserCodeWithRelations[];
 };
